Define Code Coverage

+++ src/Drivers/BaseDriver.php
@@ -1,7 +1,9 @@
-<?php namespace Seedling\Drivers;
+<?php
+namespace Seedling\Drivers;
 
 use Seedling\KeyGenerators\KeyGeneratorInterface;
 use Seedling\KeyGenerators\Sha1;
+use PDO;
 
 abstract class BaseDriver
 {
@@ -12,24 +14,34 @@ abstract class BaseDriver
      * @var array
      */
     protected $tables = array();
-    
+
+    /**
+     * @var KeyGeneratorInterface
+     */
     protected $keyGenerator;
-    
+
+    /**
+     * @var PDO
+     */
+    protected $db;
+
     /**
      * Constructor method
      *
      * @param KeyGeneratorInterface $keyGenerator
+     * @param PDO $db
      */
-    public function __construct(KeyGeneratorInterface $keyGenerator = null)
+    public function __construct(KeyGeneratorInterface $keyGenerator = null, PDO $db = null)
     {
         if (null === $keyGenerator) {
             $keyGenerator = new Sha1();
         }
         $this->keyGenerator = $keyGenerator;
+        $this->db = $db;
     }
 
     /**
-     * Truncate a table.
+     * Truncate tables
      *
      * @param array $tables The tables to truncate, null for only tables
      * that have been inserted into in this instance.
@@ -37,10 +49,14 @@ abstract class BaseDriver
      */
     public function truncate(array $tables = null)
     {
+        if (null === $this->db) {
+            return;
+        }
+
         if (null === $tables) {
             $tables = $this->tables;
         }
-        
+
         foreach ($tables as $table) {
             $this->db->query("DELETE FROM $table");
         }
@@ -57,13 +73,5 @@ abstract class BaseDriver
     protected function generateKey($value, $tableName = null)
     {
         return $this->keyGenerator->generateKey($value, $tableName);
-    }
-
-    /**
-     * Set tables that the driver to process
-     */
-    protected function setTables(array $tables)
-    {
-        
     }
 }
